Task #244250: Add two columns to store completionpercent

+++ app/src/tracking/entities/lesson-track.entity.ts
@@ -68,7 +68,7 @@ export class LessonTrack {
   @Column({ type: 'int', nullable: true, default: 0 })
   timeSpent: number;
 
-  @Column({ type: 'int', nullable: true, default: 0 })
+  @Column({ type: 'float', nullable: true, default: 0 })
   completionPercentage: number;
 
   @Column({ type: 'jsonb', nullable: true })
